requirements.txt
asgiref==3.9.1
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.0
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
